[0.3.x] fix: MEMORY_BASIC_INFORMATION64::RegionSize is ULONGLONG
git-svn-id: https://kdlibcpp.svn.codeplex.com/svn@87630 9b283d60-5439-405e-af05-b73fd8c4d996

+++ kdlib/include/kdlib/memaccess.h
@@ -16,7 +16,7 @@ bool readMemoryUnsafe( MEMOFFSET_64 offset, void* buffer, size_t length, bool ph
 bool isVaValid( MEMOFFSET_64 addr );
 bool compareMemory( MEMOFFSET_64 addr1, MEMOFFSET_64 addr2, size_t length, bool phyAddr = false );
 MEMOFFSET_64 searchMemory( MEMOFFSET_64 beginOffset, unsigned long length, const std::vector<char>& pattern );
-MEMOFFSET_64 findMemoryRegion( MEMOFFSET_64 beginOffset, MEMOFFSET_64& regionOffset, size_t& regionLength );
+MEMOFFSET_64 findMemoryRegion( MEMOFFSET_64 beginOffset, MEMOFFSET_64& regionOffset, unsigned long long &regionLength );
 kdlib::MemoryProtect getVaProtect( kdlib::MEMOFFSET_64 offset );
 
 ///////////////////////////////////////////////////////////////////////////////
